WIP: can input and start autoControl from keypad

+++ include/AGIS_Types.h
@@ -9,7 +9,7 @@ enum class infusionState_t {
   NOT_STARTED,           // when the board is powered on and no drops detected
   STARTED,               // as soon as drops are detected
   IN_PROGRESS,           // when infusion is started by user and not completed yet
-  ALARM_COMPLETED,       // when infusion has completed, i.e. infusedVolume reaches the target volume
+  ALARM_COMPLETED,       // when infusion has completed, i.e. infusedVolume_x100 reaches the target volume
   ALARM_STOPPED,         // when infusion stopped unexpectly, it's likely to have a problem
   ALARM_VOLUME_EXCEEDED  // when infusion has completed but we still detect drops
   // add more states here when needed
@@ -23,7 +23,7 @@ enum class infusionState_t {
 typedef struct _infusion_monitoring_data_handle_t {
   volatile unsigned int * numDrops_p;
   volatile unsigned int * dripRate_p;
-  volatile float * infusedVolume_p;
+  volatile unsigned int * infusedVolume_p;
   volatile unsigned long * infusedTime_p;
   infusionState_t * infusionState_p;
 } infusion_monitoring_data_handle_t;
